move logout button to navbar
